+++ ai_sapiens_interfaces/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package ai_sapiens_interfaces
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.2.1 (2026-09-01)
+------------------
+* None
+
 0.2.0 (2026-08-26)
 ------------------
 * Added the ``KeyboardInput`` message for keyboard teleoperation.